+++ Content.Server/Mindshield/FakeMindShield/FakeMindShieldImplantSystem.cs
@@ -1,14 +1,11 @@
 ﻿using Content.Shared.Implants;
-using Content.Shared.Implants.Components;
 using Content.Shared.Mindshield.Components;
 using Content.Shared.Mindshield.FakeMindShield;
-using Content.Shared.Tag;
 
 namespace Content.Server.Mindshield.FakeMindShield;
 
 public sealed class FakeMindShieldImplantSystem : SharedFakeMindShieldImplantSystem
 {
-    [Dependency] private readonly TagSystem _tag = default!;
 
     public override void Initialize()
     {
