+++ Content.Shared/Mindshield/FakeMindShield/SharedFakeMindShieldImplantSystem.cs
@@ -1,19 +1,20 @@
 ﻿using Content.Shared.Actions;
-using Content.Shared.Implants;
 using Content.Shared.Implants.Components;
 using Content.Shared.Mindshield.Components;
-using Content.Shared.Tag;
 
 namespace Content.Shared.Mindshield.FakeMindShield;
 
 public abstract class SharedFakeMindShieldImplantSystem : EntitySystem
 {
-    [Dependency] public readonly SharedActionsSystem _actionsSystem = default!;
+    [Dependency] private readonly SharedActionsSystem _actionsSystem = default!;
     public override void Initialize()
     {
         base.Initialize();
         SubscribeLocalEvent<SubdermalImplantComponent, FakeMindShieldToggleEvent>(OnFakeMindShieldToggle);
     }
+    /// <summary>
+    /// Raise the Action of a Implanted user toggling their implant to the FakeMindshieldComponent on their entity
+    /// </summary>
     private void OnFakeMindShieldToggle(EntityUid uid,
         SubdermalImplantComponent component,
         FakeMindShieldToggleEvent ev)
@@ -22,9 +23,9 @@ public abstract class SharedFakeMindShieldImplantSystem : EntitySystem
         if (component.ImplantedEntity is not { } ent)
             return;
 
-        if (TryComp<FakeMindShieldComponent>(ent, out var comp))
-        {
-            _actionsSystem.SetToggled(ev.Action, !comp.IsEnabled); // Set it to what the Mindshield component WILL be after this
-            RaiseLocalEvent(ent, ev);
-        }
-    }}
+        if (!TryComp<FakeMindShieldComponent>(ent, out var comp))
+            return;
+        _actionsSystem.SetToggled(ev.Action, !comp.IsEnabled); // Set it to what the Mindshield component WILL be after this
+        RaiseLocalEvent(ent, ev);
+    }
+}
