+++ Resources/Prototypes/Catalog/thief_toolbox_sets.yml
@@ -16,6 +16,7 @@
   - ClothingShoesChameleon
   - BarberScissors
   - ChameleonProjector
+  - FakeMindShieldImplanter
   - AgentIDCard
 
 - type: thiefBackpackSet
